fix js lint issue
