+++ .eslintrc-administration.js
@@ -4,5 +4,7 @@ module.exports = {
     ],
     parserOptions: {
         requireConfigFile: false,
+        ecmaVersion: 2020,
+        sourceType: 'module',
     },
 };
