+++ src/Resources/app/administration/src/component/nosto-entity-listing/index.js
@@ -1,3 +1,8 @@
+/**
+ * @sw-package innovation
+ */
+/* eslint-env es6 */
+
 import template from './nosto-entity-listing.html.twig';
 import './nosto-entity-listing.scss';
 
@@ -5,6 +10,8 @@ const { Component } = Shopware;
 
 Component.extend('nosto-entity-listing', 'sw-entity-listing', {
     template,
+
+emits: ['select-all-items', 'select-item'],
 
     props: {
         items: {
