+++ src/Resources/app/administration/src/component/nosto-grouped-view/index.js
@@ -1,291 +1,9 @@
-import template from './nosto-grouped-view.html.twig';
-import JobHelper from '../../util/job.helper';
-import './nosto-grouped-view.scss';
+/**
+ * @sw-package innovation
+ */
+/* eslint-env es6 */
 
-const { Component, Mixin } = Shopware;
-const { Criteria } = Shopware.Data;
+const { Component } = Shopware;
 
 /** @private */
-Component.register('nosto-grouped-view', {
-    template,
-
-    inject: [
-        'repositoryFactory',
-    ],
-
-    mixins: [
-        Mixin.getByName('notification'),
-        Mixin.getByName('nosto-scheduler-utils'),
-    ],
-
-    props: {
-        jobTypes: {
-            type: Array,
-            required: false,
-            default: () => [],
-        },
-        groupCreationDate: {
-            type: Object,
-            required: false,
-            default: () => {},
-        },
-        sortType: {
-            type: String,
-            required: true,
-            default: () => 'status',
-        },
-    },
-
-    data() {
-        return {
-            groupedItems: [],
-            isLoading: false,
-            showJobInfoModal: false,
-            showJobSubsModal: false,
-            currentJobID: null,
-            showMessagesModal: false,
-            currentJobMessages: null,
-            page: 1,
-            limit: 25,
-        };
-    },
-
-    computed: {
-        jobRepository() {
-            return this.repositoryFactory.create('nosto_scheduler_job');
-        },
-
-        messageRepository() {
-            return this.repositoryFactory.create('nosto_scheduler_job_message');
-        },
-
-        jobMessagesColumns() {
-            return [
-                {
-                    property: 'message',
-                    dataIndex: 'message',
-                    label: this.$tc('job-listing.page.listing.grid.column.message'),
-                    allowResize: false,
-                    align: 'left',
-                    width: '90px',
-                },
-            ];
-        },
-
-        columns() {
-            return [
-                {
-                    property: 'name',
-                    label: this.$tc('job-listing.page.listing.grid.column.name'),
-                    allowResize: true,
-                    width: '500px',
-                },
-                {
-                    property: 'status',
-                    label: this.$tc('job-listing.page.listing.grid.column.status'),
-                    allowResize: true,
-                    width: '150px',
-                },
-                {
-                    property: 'startedAt',
-                    label: this.$tc('job-listing.page.listing.grid.column.started-at'),
-                    allowResize: true,
-                    width: '170px',
-                },
-                {
-                    property: 'finishedAt',
-                    label: this.$tc('job-listing.page.listing.grid.column.finished-at'),
-                    allowResize: true,
-                    width: '170px',
-                },
-                {
-                    property: 'createdAt',
-                    label: this.$tc('job-listing.page.listing.grid.column.created-at'),
-                    allowResize: true,
-                    width: '170px',
-                },
-                {
-                    property: 'subJobs',
-                    label: this.$tc('job-listing.page.listing.grid.column.child-jobs'),
-                    allowResize: true,
-                    width: '250px',
-                    visible: true,
-                    sortable: false,
-                },
-                {
-                    property: 'messages',
-                    label: this.$tc('job-listing.page.listing.grid.column.messages'),
-                    allowResize: true,
-                    width: '350px',
-                    visible: true,
-                    sortable: false,
-                },
-            ];
-        },
-    },
-
-    watch: {
-        sortType() {
-            this.groupedItems = [];
-            this.initGroupedView();
-        },
-        groupCreationDate: {
-            handler() {
-                this.groupedItems = [];
-                this.initGroupedView();
-            },
-            deep: true,
-        },
-    },
-
-    created() {
-        this.initGroupedView();
-    },
-
-    beforeDestroy() {
-        clearInterval(this.reloadInterval);
-    },
-
-    methods: {
-        initGroupedView() {
-            this.isLoading = true;
-            const criteria = new Criteria();
-            criteria.addFilter(Criteria.equals('parentId', null));
-            criteria.addSorting(Criteria.sort('createdAt', 'DESC', false));
-            criteria.addAssociation('messages');
-            criteria.addAssociation('subJobs');
-            criteria.limit = null;
-
-            if (this.jobTypes !== []) {
-                criteria.addFilter(Criteria.equalsAny('type', this.jobTypes));
-            }
-
-            if (this.groupCreationDate.fromDate) {
-                criteria.addFilter(Criteria.range('createdAt', { gte: this.groupCreationDate.fromDate }));
-            }
-
-            if (this.groupCreationDate.toDate) {
-                criteria.addFilter(Criteria.range('createdAt', { lte: this.groupCreationDate.toDate }));
-            }
-
-            return this.jobRepository.search(criteria, Shopware.Context.api).then(items => {
-                this.sortJobs(items);
-            });
-        },
-
-        sortJobs(items) {
-            this.groupedItemsTypes = [];
-            this.groupedItems = [];
-            items.forEach((item) => {
-                const index = this.groupedItemsTypes.findIndex(e => e.title === item[this.sortType]);
-                if (index === -1) {
-                    this.groupedItemsTypes.push({
-                        title: item[this.sortType],
-                    });
-                }
-            });
-
-            this.getJobsByType(this.groupedItemsTypes);
-        },
-
-        getJobsByType(types) {
-            types.forEach((type) => {
-                const criteria = new Criteria(this.page, this.limit);
-                criteria.addFilter(Criteria.equals('parentId', null));
-                criteria.addSorting(Criteria.sort('createdAt', 'DESC', false));
-                criteria.addAssociation('messages');
-                criteria.addAssociation('subJobs');
-
-                if (this.groupCreationDate.fromDate) {
-                    criteria.addFilter(Criteria.range('createdAt', { gte: this.groupCreationDate.fromDate }));
-                }
-
-                if (this.groupCreationDate.toDate) {
-                    criteria.addFilter(Criteria.range('createdAt', { lte: this.groupCreationDate.toDate }));
-                }
-
-                if (this.jobTypes !== []) {
-                    criteria.addFilter(Criteria.equalsAny('type', this.jobTypes));
-                }
-
-                if (this.sortType === 'status') {
-                    criteria.addFilter(Criteria.equals('status', type.title));
-                }
-
-                if (this.sortType === 'type') {
-                    criteria.addFilter(Criteria.equals('type', type.title));
-                }
-
-                this.jobRepository.search(criteria, Shopware.Context.api).then(items => {
-                    const groupType = this.sortType === 'status' ? type.title.toUpperCase() : items[0].name;
-                    const groupTitle = this.sortType === 'status'
-                        ? this.$tc(`job-listing.page.listing.grid.job-status.${type.title}`)
-                        : items[0].name;
-                    this.groupedItems.push({
-                        title: groupTitle,
-                        type: groupType,
-                        items: JobHelper.sortMessages(items),
-                    });
-                });
-            });
-
-            this.isLoading = false;
-
-            return this.groupedItems;
-        },
-
-        getMessagesCount(job, type) {
-            return job.messages.filter((item) => {
-                return item.type === `${type}-message`;
-            }).length;
-        },
-
-        getChildrenCount(job, type) {
-            return job.subJobs.filter((item) => {
-                return item.status === type;
-            }).length;
-        },
-
-        onRefresh() {
-            this.initGroupedView();
-        },
-
-        canDelete(item) {
-            return ['error', 'succeed'].indexOf(item.status) !== -1;
-        },
-
-        onDeleteJob(jobId) {
-            this.jobRepository.delete(jobId, Shopware.Context.api).then(() => {
-                this.updateList();
-            });
-        },
-
-        rescheduleJob(jobId) {
-            this.NostoRescheduleService.rescheduleJob(jobId).then(() => {
-                this.createNotificationSuccess({
-                    message: 'Job has been rescheduled successfully.',
-                });
-                this.updateList();
-            }).catch(() => {
-                this.createNotificationError({
-                    message: 'Unable reschedule job.',
-                });
-            });
-        },
-
-        showJobInfo(jobId) {
-            this.currentJobID = jobId;
-            this.showJobInfoModal = true;
-        },
-
-        showSubJobs(jobId) {
-            this.currentJobID = jobId;
-            this.showJobSubsModal = true;
-        },
-
-        showJobMessages(job) {
-            this.currentJobMessages = job.messages;
-            this.showMessagesModal = true;
-        },
-    },
-});
+Component.register('nosto-grouped-view', () => import('./nosto-grouped-view.component'));
