+++ src/Resources/app/administration/src/component/nosto-job-info/index.js
@@ -1,51 +1,9 @@
-import template from './nosto-job-info.html.twig';
+/**
+ * @sw-package innovation
+ */
+/* eslint-env es6 */
 
-const { Component, Mixin } = Shopware;
-const { Criteria } = Shopware.Data;
+const { Component } = Shopware;
 
 /** @private */
-Component.register('nosto-job-info', {
-    template,
-
-    inject: [
-        'NostoRescheduleService',
-        'repositoryFactory',
-    ],
-
-    mixins: [
-        Mixin.getByName('notification'),
-        Mixin.getByName('nosto-scheduler-utils'),
-    ],
-
-    props: {
-        jobId: {
-            type: String,
-            required: false,
-            default: null,
-        },
-    },
-
-    data() {
-        return {
-            jobItem: null,
-        };
-    },
-
-    computed: {
-        jobRepository() {
-            return this.repositoryFactory.create('nosto_scheduler_job');
-        },
-    },
-
-    created() {
-        this.initPageData();
-    },
-
-    methods: {
-        initPageData() {
-            this.jobRepository.get(this.jobId, Shopware.Context.api, new Criteria()).then(jobItem => {
-                this.jobItem = jobItem;
-            });
-        },
-    },
-});
+Component.register('nosto-job-info', () => import('./nosto-job-info.component'));
