+++ src/Resources/app/administration/src/component/nosto-job-status-badge/index.js
@@ -1,33 +1,10 @@
-import template from './nosto-job-status-badge.html.twig';
+/**
+ * @sw-package innovation
+ */
+/* eslint-env es6 */
 
 const { Component } = Shopware;
 
+
 /** @private */
-Component.register('nosto-job-status-badge', {
-    template,
-
-    props: {
-        status: {
-            type: String,
-            required: true,
-        },
-    },
-
-    computed: {
-        additionalClass() {
-            return this.status === 'running' ? '--pulse' : '';
-        },
-
-        variant() {
-            switch (this.status) {
-                case 'error':
-                    return 'error';
-                case 'succeed':
-                case 'running':
-                    return 'success';
-                default:
-                    return '';
-            }
-        },
-    },
-});
+Component.register('nosto-job-status-badge', () => import('./nosto-job-status-badge.component'));
